vaccum-registry: separate numeric and non-numeric tokens in tag
comparing integers and strings fails with
`ArgumentError (comparison of Integer with String failed)`, thus in
order to correctly compare tags that are structurally different, we need
to explicitly define the comparison order. We're doing this by
transforming numerics into `[0, <int_value>]` and strings into `[1,
<str_value>]`; this ensures that comparisons between ints and strings
are always decided on the first tuple value (with numeric values
considered smaller than non-numeric values) and there's never a
comparison betwen the int and the string value.